requirements.txt
asgiref==3.8.1
boto3==1.35.63
botocore==1.35.63
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-ses==4.2.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.2
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
